New ObselCollection Class : this class is an obsel container.
What : It implements IList so it could be use as a DataProvider. It works with an Obsel Vector underneath (for now on, this vector is accessible with _obsels public attribute) 

Why : because it is based on a Vector, it make it a lot faster than ArrayCollection or Array to access or create item (and then a lot faster to filter a collections).
Also this container should be use in applications in order to handle amount of various obsels.

TODO : that is a first working version. There is still a lot todo to make it cleaner. It should implements completely IList, a lot a functions are still empty. Also the attribute _obsels has to be renamed or replaced by a function. Some utilitary function should also written, like filtering functions.